refactor(#792): decompose block_editor_window_lifecycle::wire (536-line fn) by responsibility
Orchestrator + 4 helpers: wire_model_selection, wire_drawer_toggle_save (base) +
wire_block_delete, wire_plugin_info_close (block_editor_window_delete.rs sibling).
Same closures/captures via &Ctx. File 678->466, off debt. adapter-gui lib compiles clean.
