+++ crates/adapter-gui/src/lib.rs
@@ -18,6 +18,7 @@ mod block_drawer_close_wiring;
 mod block_drawer_save_delete_wiring;
 pub mod block_editor_param_tabs;
 mod block_editor_window_lifecycle;
+mod block_editor_window_delete;
 mod block_editor_window_params;
 mod block_editor_window_setup;
 mod block_editor_window_wiring;
